refactor : 해당 클래스들 post/application/command 폴더 밑으로 이동
