+++ src/main/java/com/kakaotech/team14backend/post/application/command/FindLikeStatusService.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.post.application.service;
+package com.kakaotech.team14backend.post.application.command;
 
 import com.kakaotech.team14backend.post.domain.PostLike;
 import com.kakaotech.team14backend.post.infrastructure.PostLikeRepository;
