+++ src/main/java/com/kakaotech/team14backend/post/application/command/FindPopularPosts.java
@@ -1,13 +1,14 @@
-package com.kakaotech.team14backend.post.application;
+package com.kakaotech.team14backend.post.application.command;
 
 import com.kakaotech.team14backend.common.RedisKey;
-import com.kakaotech.team14backend.outer.post.dto.GetIncompletePopularPostDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListRequestDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListResponseDTO;
+import com.kakaotech.team14backend.post.application.PostMapper;
 import com.kakaotech.team14backend.post.domain.LevelIndexes;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.domain.PostRandomFetcher;
 import com.kakaotech.team14backend.post.domain.RandomIndexes;
+import com.kakaotech.team14backend.post.dto.GetIncompletePopularPostDTO;
+import com.kakaotech.team14backend.post.dto.GetPopularPostListRequestDTO;
+import com.kakaotech.team14backend.post.dto.GetPopularPostListResponseDTO;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
